fix: token expiry error

+++ MySignalsApp/utils.py
@@ -3,7 +3,7 @@ from MySignalsApp.models.base import get_uuid
 from MySignalsApp.models.signals import Signal
 from MySignalsApp.models.placed_signals import PlacedSignals
 from MySignalsApp.models.user_tokens import UserTokens
-from datetime import datetime, timedelta
+from datetime import datetime
 from flask import current_app, url_for, render_template
 from MySignalsApp import db, mail
 from flask_mail import Message
@@ -46,9 +46,9 @@ def query_paginate_filtered(table, page, **kwargs):
 
 
 # token helpers
-def get_reset_token(user, expires=datetime.utcnow() + timedelta(hours=1)):
+def get_reset_token(user):
     token = get_uuid()
-    token_data = UserTokens(user_id=user.id, token=token, expiration=expires)
+    token_data = UserTokens(user_id=user.id, token=token)
     token_data.insert()
     return token
 
@@ -58,14 +58,13 @@ def verify_reset_token(user_table, token):
         token_data = query_one_filtered(UserTokens, token=token)
         if not token_data:
             return None
-
-        if datetime.utcnow() <= token_data.expiration:
-            user = query_one_filtered(user_table, id=token_data.user_id)
+        user = query_one_filtered(user_table, id=token_data.user_id)
+        if token_data.expiration >= datetime.utcnow():
             token_data.delete()
             return user
-        else:
-            token_data.delete()
-            return None
+
+        token_data.delete() if user.is_active else user.delete()
+        return None
     except Exception as e:
         current_app.log_exception(e)
         raise UtilError("Internal server error", 500, "It's not you it's us")
